refactor(analytics): cut over to native Python Hatchet
