--- apps/hatchet-worker-analytics/tsconfig.json
@@ -1,26 +1,0 @@
-{
-  "include": ["./src/**/*", "./scripts/**/*"],
-  "compilerOptions": {
-    "baseUrl": ".",
-    /* Base Options: */
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "target": "es2022",
-    "allowJs": true,
-    "resolveJsonModule": true,
-    "moduleDetection": "force",
-    "isolatedModules": true,
-    "verbatimModuleSyntax": true,
-    /* Strictness */
-    "strict": true,
-    "noUncheckedIndexedAccess": true,
-    "noImplicitOverride": true,
-    /* If transpiling with TypeScript: */
-    "module": "NodeNext",
-    "outDir": "dist",
-    "sourceMap": true,
-    "tsBuildInfoFile": "dist/tsconfig.tsbuildinfo",
-    /* If your code doesn't run in the DOM: */
-    "lib": ["es2022"]
-  }
-}
