+++ deploy/charts/klicker-uzh-v3/templates/cm-hatchet-workers.yaml
@@ -8,6 +8,8 @@ metadata:
 data:
   HATCHET_CLIENT_TLS_STRATEGY: {{ .Values.hatchet.client.tlsStrategy | quote }}
   HATCHET_API_URL: {{ .Values.hatchet.client.apiUrl | quote }}
+  # Days after Course.endDate before the scanner emits a finalization event.
+  ANALYTICS_FINALIZE_GRACE_DAYS: '7'
 ---
 apiVersion: v1
 kind: ConfigMap
@@ -39,11 +41,5 @@ metadata:
 data:
   HATCHET_CLIENT_TLS_STRATEGY: {{ .Values.hatchet.client.tlsStrategy | quote }}
   HATCHET_API_URL: {{ .Values.hatchet.client.apiUrl | quote }}
-  # ANALYTICS_CWD matches the WORKDIR baked into the analytics worker image.
-  # Override in a per-env ConfigMap if the layout ever changes.
-  ANALYTICS_CWD: /app/apps/analytics
-  ANALYTICS_RUNNER_CMD: uv
-  ANALYTICS_RUNNER_ARGS: 'run python'
-  # Days after Course.endDate before the scanner marks a course as FINALIZING.
-  # Keeps the grace window generous for late student responses / chat messages.
-  ANALYTICS_FINALIZE_GRACE_DAYS: '7'
+  HATCHET_CLIENT_WORKER_HEALTHCHECK_ENABLED: 'true'
+  HATCHET_CLIENT_WORKER_HEALTHCHECK_PORT: '8001'
