+++ deploy/env-uzh-prd/values.yaml
@@ -134,6 +134,14 @@ hatchet:
           value: reserved
           effect: NoSchedule
 
+    analytics:
+      replicaCount: 1
+      priorityClassName: production-workload
+      image:
+        repository: ghcr.io/uzh-bf/klicker-uzh/hatchet-worker-analytics-arm
+        pullPolicy: Always
+        tag: v3.4.0-alpha.62
+
 auth:
   priorityClassName: production-workload
