+++ packages/graphql/src/index.ts
@@ -38,7 +38,6 @@ import './schema/subscription.js'
 //   })
 // }
 
-import { handleRunAnalyticsScript } from './services/analyticsRecompute.js'
 import {
   handleEndExpiredGroupActivity,
   handleFinalRandomGroupAssignments,
@@ -83,5 +82,4 @@ export const handlers: HatchetHandlers = {
   handlePublishScheduledMicroLearning,
   handleStandardLiveQuizBlockClosureAggregation,
   handleAssessmentLiveQuizBlockClosureAggregation,
-  handleRunAnalyticsScript,
 }
