+++ turbo.json
@@ -3,11 +3,8 @@
   "globalEnv": [
     "ANALYTICS_ALLOW_FULL",
     "ANALYTICS_COURSE_IDS",
-    "ANALYTICS_CWD",
     "ANALYTICS_FINALIZE_GRACE_DAYS",
     "ANALYTICS_MODE",
-    "ANALYTICS_RUNNER_ARGS",
-    "ANALYTICS_RUNNER_CMD",
     "ANALYTICS_WINDOW_SINCE",
     "API_DOMAIN",
     "APP_CONTROL_DOMAIN",
